moving dashboard from containers to components
