+++ app/src/components/index.js
@@ -1,2 +1,3 @@
 export Navigation from './Navigation/Navigation';
 export Login from './Login/Login';
+export Dashboard from '.Dashboard/Dashboard';
