+++ app/src/containers/index.js
@@ -6,7 +6,7 @@ export Home from './Home/Home';
 export Main from './Main/Main';
 export Maps from './Maps/Maps';
 export Test from './Test/Test';
-export Dashboard from './Dashboard/Dashboard';
+// export Dashboard from './Dashboard/Dashboard';
 export EventsList from './Events/EventsList';
 export EventDetails from './Event/EventDetails';
 export Guest from './Guest/sampleGuest';
